+++ Pandora/Paradigm/Structure/Interface/Set.hs
@@ -18,8 +18,9 @@ import Pandora.Paradigm.Primary.Functor.Predicate (Predicate, equate)
 import Pandora.Paradigm.Primary.Object.Boolean (Boolean)
 import Pandora.Paradigm.Primary.Object.Numerator (Numerator (Zero))
 import Pandora.Paradigm.Schemes.T_U (type (<:.:>))
+import Pandora.Paradigm.Inventory.Ability.Modifiable (modify_)
 import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing), Morph (Find), find)
-import Pandora.Paradigm.Inventory.Some.State (State, modify)
+import Pandora.Paradigm.Inventory.Some.State (State)
 import Pandora.Paradigm.Controlflow.Effect (run, (!))
 
 type Set t f a = (Traversable (->) (->) t, Setoid a, Setoid (t a), Morphable (Find f) t)
@@ -28,4 +29,4 @@ subset :: forall t f a . (Set t f a, Morphing (Find f) t ~ (Predicate <:.:> Mayb
 subset = Convergence ! \s ss -> Nothing != (find @f @t @Maybe % s) . equate <<- ss
 
 cardinality :: Traversable (->) (->) t => t a -> Numerator
-cardinality s = attached . run @(->) @(State _) % Zero ! constant (modify @Numerator (+ one)) <<- s
+cardinality s = attached . run @(->) @(State _) % Zero ! constant (modify_ @State (+ one)) <<- s