+++ CHANGELOG.md
@@ -660,3 +660,4 @@
 * Define `Ability` umbrella module (like in `Structure`)
 * Define experimenal `Viewable`, `Replaceable` and `Modifiable` typeclasses
 * Define `Some` umbrella module (like in `Structure`)
+* Remove `modify` method of `State` effect